java italian iter_6

+++ src/main/java/org/636767031a6d9265ec0178e6/636767031a6d9265ec0178e6.java
@@ -19,7 +19,7 @@ public class ArrayUtils {
   
   final byte[] result = new byte[array.length];
   for (int i = 0; i < array.length; i++) {
-  result[i] = array[i].byteValue();
+  result[i] = array[i]; // Throws NPE if array[i] is null
   }
   return result;
   }
